rename umap method and remove duplicate code
