+++ src/vame/analysis/community_analysis.py
@@ -24,7 +24,7 @@ from vame.util.data_manipulation import consecutive
 from typing import List, Tuple
 from vame.schemas.states import save_state, CommunityFunctionSchema
 from vame.logging.logger import VameLogger
-from vame.analysis.umap_visualization import umap_vis_community_labels
+from vame.analysis.umap import umap_vis_community_labels, umap_embedding
 
 
 logger_config = VameLogger(__name__)
@@ -414,37 +414,6 @@ def get_cohort_community_labels(
     community_labels_all.append(community_labels)
 
     return community_labels_all
-
-
-def umap_embedding(cfg: dict, file: str, model_name: str, n_cluster: int, parametrization: str) -> np.ndarray:
-    """Perform UMAP embedding for given file and parameters.
-
-    Args:
-        cfg (dict): Configuration parameters.
-        file (str): File path.
-        model_name (str): Model name.
-        n_cluster (int): Number of clusters.
-        parametrization (str): parametrization.
-
-    Returns:
-        np.ndarray: UMAP embedding.
-    """
-    reducer = umap.UMAP(n_components=2, min_dist=cfg['min_dist'], n_neighbors=cfg['n_neighbors'],
-                        random_state=cfg['random_state'])
-
-    logger.info("UMAP calculation for file %s" %file)
-
-    folder = os.path.join(cfg['project_path'],"results",file,model_name, parametrization +'-'+str(n_cluster),"")
-    latent_vector = np.load(os.path.join(folder,'latent_vector_'+file+'.npy'))
-
-    num_points = cfg['num_points']
-    if num_points > latent_vector.shape[0]:
-        num_points = latent_vector.shape[0]
-    logger.info("Embedding %d data points.." %num_points)
-
-    embed = reducer.fit_transform(latent_vector[:num_points,:])
-
-    return embed
 
 
 @save_state(model=CommunityFunctionSchema)
